+++ sensor_interface/scripts/ms5837_interface.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+
+import rospy
+
+class Ms5837InterfaceNode(object):
+    def __init__(self):
+        rospy.init_node('pressure_node')
+        #Stuff
+
+if __name__ == '__main__':
+    try:
+        pressure_node = Ms5837InterfaceNode()
+        rospy.spin()
+    except rospy.ROSInterruptException:
+        pass
